Add logic to record whether CBT was computed for a circuit
A new field, cbt_set, is added for every circuit; this is set to
false at the beginning of a measurement, and gets set to True
after a BUILDTIMEOUT_SET COMPUTED event. Each
BUILDTIMEOUT_SET RESET event resets the field to False.
Any measurements that specify `--drop-guards` will use this
field to exclude circuits built when CBT was not computed
from the daily analysis by default.

Analysis files for these measurements are also tagged by default with
the 'exclude_cbt' filter, which is then used by the visualize command to
only visualize results that used non-excluded circuits.